fastapi/requirements.txt
# Langchain libraries
langchain-core
langchain-openai
langchain-anthropic
langchain-community
langchain-ollama
langchain-experimental

# Other libraries
python-dotenv

# FastAPI libraries
fastapi
uvicorn
requests    

# SQLModel for database
sqlmodel




